backend/requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
docxtpl==0.16.7
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
MarkupSafe==3.0.2
openai==1.65.4
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.2
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
